refactor: merge core/scripts/lib/log-parser into core/lib/log
